Massive refactoring of IO handling
Switch around WindowsIOManager so that the main thread runs select(),
and the child thread checks the IOCP. This is much much easier to work
with, because you can queue events to the IOCP from any thread. It's
also way more efficient than the call_soon stuff I was doing before.

And now that we have wait_socket_readable everywhere and it's the most
efficient thing, using a socketpair() for the call_soon wakeup channel
everywhere makes more sense than using IOCP on Windows and a pipe on
!Windows. So factor that out into common code. And then on Windows we
can re-use this code for the wakeup channel between the IOCP thread
and the main thread.