MultiError checkpoint
I'm running into a tangle of nasty problems here, and might need to
fundamentally rethink things, so it's a good time for a checkpoint.

I originally had the "handler" passed to MultiError.{a,}{filter,catch}
be a generator sorta like a @contextmanager, where it would throw in
each of the leaf errors. The problem with this is that it mutates the
stack of every exception, every time, which totally messes up attempts
to keep the tracebacks non-redundant. Is it even worth trying to keep
them non-redundant? Maybe we should just flatten everything out and
call it a day...? ughhh.

__context__ and __cause__ are also being really nasty. Partly because
I'm fighting Python who wants to keep adding them on. For example, the
nursery cleanup code, if entering with an exception set, ends up
attaching that context to every incoming exception, which makes no
sense at all! --I'm not sure this is the full extent of it either!  --
this checkpoint has some overbroad __suppress_context__=True
assignments, with print calls to remind us to remove them. Maybe we
should not unwrap() at all? and shield immediately when calling
cancel?

But also because, what do we do when snapping a MultiError smaller?
Can a MultiError ever legitimately have a cause/context attached? I
think so in weird cases like:

except:
    with open_nursery() as nursery:
        ...

so when this happens, do we push down the __cause__/__context__ onto
the children, along with the traceback?