Work around flakiness in test_ki_wakes_us_up
This is being flaky on Windows, almost certainly due to a bug in
CPython 3.5.{0,1,2,3} and 3.6.{0,1}:
    https://bugs.python.org/issue30038
More details in gh-119.

I've submitted (what I think is) a fix to CPython, but in the mean
time the random appveyor failures are pretty annoying and I don't see
any way to fix them except wait for CPython to make a new release, so
as a temporary measure this commit adds some retry logic to this
test.

This is an ugly hack and should be removed as soon as possible.
(Unfortunately that probably won't be until after we can drop 3.6
support...)