New argument to trio.run: restrict_keyboard_interrupt_to_checkpoints
There are two main strategies for handling control-C:

- Accept that a KeyboardInterrupt will be raised "asynchronously" (in
  the sense of "asynchronous exceptions", nothing to do with
  async/await). It's essentially impossible to write fully correct
  code under this model (see many rants by Java folks), but it does
  something more-or-less reasonable 99% of the time, and it's what
  Python developers expect.

- Use some sort of explicit handler for KeyboardInterrupt, and only
  deal with them at carefully chosen places in your code where you
  know it's safe to do so. This makes it possible to write fully
  correct code, but doesn't make it easy; such code tends to be
  difficult to write and fragile. It's not *as* bad in trio because we
  have nice abstractions like cancellation and trio.catch_signals, but
  there are still gotchas (e.g. catch_signals raises an error if you
  try to call it outside the main thread; that's pretty easy to miss
  in testing!).

I realized that it's pretty trivial for us to provide a nice third
option combines many of the advantages of these two approaches: we
already have the machinery to route KeyboardInterrupt through the
cancellation machinery when necessary, so this patch adds a flag that
lets the user request that this happen all the time. Flipping on this
flag gives you most of the advantages of the "explicit handler" model,
without having to write any code.