[doc] Emphasize that parenting is a full-time job in the tutorial
Some early user feedback suggests that this is possible to
miss (gh-135), and it's important and kinda subtle (it involves an
invisible function call!). So let's make sure to point it out
explicitly the very first time we see a call to open_nursery, and
reinforce it as we go.