Allow TLS <= v1.2, but not 1.3, in the renegotiation tests
The problem is we want to require < TLS 1.3, because those are the
versions that have renegotiation, but there isn't a trivial way to say
this in (Py)OpenSSL (or at least, not on any (Py)OpenSSL we currently
have access to). We previously did this by forcing TLS 1.2 only, but
the CI tests are breaking on MacOS, because CPython 3.5 on MacOS uses
such an ancient OpenSSL that it can't speak TLS 1.2 to the test
harness.

Some discussion:
https://github.com/pyca/pyopenssl/issues/624