Fix test broken by async_generator 1.7 release
In v1.7, async_generator.isasyncgenfunction became pickier, to match
inspect.isasyncgenfunction and because this was needed to fix the trio
docs. Apparently this bit of trio's test suite was relying on the old
laxity...

In general we could also make acontextmanager more lenient, and that
might make sense if it were a public API, but for now the extra
pickiness seems fine.