Add contributing doc, and switch to WeAllJS code of conduct
We obviously need a contributing doc, and the WeAllJS code of conduct
is excellent. These are together because part of the idea here is to
move these into the docs, and have that act as a single standard
location that all the python-trio projects can point to, instead of
copying them into every repo.