Link to SSLContext.sni_callback, instead of set_servername_callback
SSLContext.sni_callback is a better replacement for
set_servername_callback (that was added in part because of my whining
about trio bug gh-11). Let's link to it instead.

This also fixes a silly issue where we were using :meth: to link to
SSLContext.set_servername_callback, but in the 3.7 docs that's
documented as being an attribute. Even though it's a method. Oh well,
whatever, if we don't link to it we don't have to care.