Add socket.bind to the list of async socket methods
I was recently using trio for a project to make a UDP listening
socket, and I didn't know to await on `socket.bind`. The warning that
I got about the coroutine never being awaited tipped me off, but it
would still be nice to have this in the docs.

I also wrote up about how to build the docs locally. I haven't
contributed to a lot of well-documented Python projects, so this was
instructive for me.