Add fail-fast: false to Github Actions config
By default, as soon as a single job in a group fails, then GA will
cancel all the other jobs. This is kind of confusing, e.g. I was just
looking at a PR that had a Python 3.5-specific failure, but that
wasn't immediately obvious, because GA only ran the 3.5 tests and then
skipped all the others. So let's disable that behavior.