Remove constraint on typed-ast requirement
Now that Dependabot is using 3.8, a requirement only applicable below 3.8 is completely ignored. But black pulls in typed-ast unconditionally, and pip-tools isn't smart enough to propagate environment constraints along dependency edges, so we wind up with typed-ast as an unconditional dependency which makes pypy builds fail. Constraining it in requirements.in based on only the interpreter, not also the Python version, resolves this (tested by running pip-compile on 3.8 directly).